correlateai-enterprise/requirements.txt
"""
CorrelateAI Enterprise - Requirements
Dependencies for the AI agent framework and correlation analysis system
"""

# Core Azure AI and database dependencies
azure-ai-projects>=1.0.0b1
azure-identity>=1.15.0
azure-core>=1.29.0

# Database connectivity
pyodbc>=5.0.0
sqlalchemy>=2.0.0
pandas>=2.1.0
numpy>=1.25.0

# Machine learning and statistics
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Web framework (for dashboard)
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0

# Frontend dependencies (React/TypeScript)
# Note: These would be managed separately via npm/yarn
# react>=18.0.0
# typescript>=5.0.0
# @types/react>=18.0.0

# Async and HTTP
aiohttp>=3.9.0
httpx>=0.25.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging and monitoring
structlog>=23.2.0
azure-monitor-opentelemetry>=1.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development tools
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0